+++ astropy/table/table.py
@@ -236,7 +236,7 @@ class Table:
     TableColumns = TableColumns
     TableFormatter = TableFormatter
 
-    def as_array(self, keep_byteorder=False, key_values=[]):
+    def as_array(self, keep_byteorder=False, names=None):
         """
         Return a new copy of the table in the form of a structured np.ndarray or
         np.ma.MaskedArray object (as appropriate).
@@ -248,10 +248,9 @@ class Table:
             order.  However, if this option is `True` this preserves the
             byte order of all columns (if any are non-native).
 
-        key_values : list, optional
-            By default the returned array contains all columns.
-            However, if this list contains some values of columns then
-            the returned array is sliced according to it.
+        names : list, optional:
+            List of column names to include for returned structured array.
+            Default is to include all table columns.
 
         Returns
         -------
@@ -266,12 +265,8 @@ class Table:
 
         dtype = []
 
-        for key in key_values:
-            if key not in self.columns:
-                print('Warning! Key does not exist')
-
-        if key_values != []:
-            cols = self[key_values].columns.values()
+        if names != None:
+            cols = self[names].columns.values()
 
         else:
             cols = self.columns.values()